# PERF: Rolling._apply