@mroeschke 2021-08-23
There's an `_apply_series` which looks like a 1D case of `_apply_blockwise`; can that be refactored into `_apply_blockwise`?

`_apply_blockwise` maybe should be renamed to `_apply_columnwise`