@jbrockmendel 2021-08-23
Updated timings:

```
from asv_bench.benchmarks.rolling import *

self = Engine()
self.setup("DataFrame", "float", sum, "cython", "sum")
self.time_rolling_methods("DataFrame", "float", sum, "cython", "sum")

%timeit self.time_rolling_methods("DataFrame", "float", sum, "cython", "sum")
209 µs ± 8.64 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
147 µs ± 3.32 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR, consolidate=True
133 µs ± 1.95 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR, consolidate=False
```

I'm OK with consolidating at that cost.