@jbrockmendel 2021-08-24
> Does this deserve a whatsnew note?

either way i guess, i can add something in the next "assorted follow-ups" PR